+++ services/key-manager/provider_test.go
@@ -22,7 +22,7 @@ func TestNewInternalConfig(t *testing.T) {
 						EntryPoints: []string{http.DefaultHTTPAppEntryPoint},
 						Service:     "ethereum",
 						Priority:    math.MaxInt32,
-						Rule:        "PathPrefix(`/ethereum/accounts`)",
+						Rule:        "PathPrefix(`/ethereum`)",
 						Middlewares: []string{"base@logger-base"},
 					},
 				},